requirements.txt
flask
ultralytics
fpdf2
pillow